fixing test

+++ datafusion/core/src/scalar.rs
@@ -644,7 +644,7 @@ mod tests {
         assert_eq!(
             List(
                 Some(vec![Int64(Some(1)), Int64(Some(5))]),
-                Box::new(DataType::Int32),
+                Box::new(DataType::Int64),
             )
             .partial_cmp(&List(
                 Some(vec![Int32(Some(1)), Int32(Some(5))]),
